renamed subpackage ssr to training
