+++ README.md
@@ -39,8 +39,9 @@ from mhn.full_state_space import Likelihood, ModelConstruction, RegularizedOptim
 ```
 You can train an MHN using state-space restriction. The corresponding functions
 can be imported with
+
 ```python
-from mhn.ssr import state_space_restriction, state_containers
+from mhn.training import state_space_restriction, state_containers
 ```
 Training a new MHN can be as simple as writing the following few lines of code:
