+++ doc/mhn.rst
@@ -38,16 +38,16 @@ The *full_state_space* subpackage
 
     View documentation of full_state_space <full_state_space/mhn.full_state_space>
 
-The *ssr* subpackage
+The *training* subpackage
 --------------------
 
-.. automodule:: mhn.ssr
+.. automodule:: mhn.training
     :noindex:
 
 .. toctree::
     :maxdepth: 1
 
-    View documentation of ssr <ssr/mhn.ssr>
+    View documentation of training <training/mhn.training>
 
 The *omega_mhn* subpackage
 --------------------------
