+++ doc/training/state_containers.rst
@@ -3,7 +3,7 @@
 The *state_containers* submodule
 --------------------------------
 
-.. automodule:: mhn.ssr.state_containers
+.. automodule:: mhn.training.state_containers
    :members:
    :undoc-members:
    :inherited-members:
