+++ doc/training/state_space_restriction.rst
@@ -3,18 +3,18 @@
 The *state_space_restriction* submodule
 ---------------------------------------
 
-.. automodule:: mhn.ssr.state_space_restriction
+.. automodule:: mhn.training.state_space_restriction
    :members:
    :undoc-members:
 
-.. data:: mhn.ssr.state_space_restriction.CUDA_AVAILABLE
+.. data:: mhn.training.state_space_restriction.CUDA_AVAILABLE
 
     constant returned by :code:`cuda_available()` if CUDA functions are available
 
-.. data:: mhn.ssr.state_space_restriction.CUDA_NOT_AVAILABLE
+.. data:: mhn.training.state_space_restriction.CUDA_NOT_AVAILABLE
 
     constant returned by :code:`cuda_available()` if CUDA functions are not available
 
-.. data:: mhn.ssr.state_space_restriction.CUDA_NOT_FUNCTIONAL
+.. data:: mhn.training.state_space_restriction.CUDA_NOT_FUNCTIONAL
 
     constant returned by :code:`cuda_available()` if CUDA functions don't work as expected, in this case check your CUDA drivers
