+++ mhn/__init__.py
@@ -1,11 +1,11 @@
 from . import optimizers
 from . import model
-from . import ssr
+from . import training
 from . import full_state_space
 from . import omega_mhn
 
-from .ssr.state_space_restriction import cuda_available
-from .ssr.state_space_restriction import CUDA_AVAILABLE, CUDA_NOT_AVAILABLE, CUDA_NOT_FUNCTIONAL
+from .training.state_space_restriction import cuda_available
+from .training.state_space_restriction import CUDA_AVAILABLE, CUDA_NOT_AVAILABLE, CUDA_NOT_FUNCTIONAL
 
 
 def _get_metadata():
