+++ mhn/model.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 from numpy.core.multiarray import array as array
 
 from .full_state_space import Likelihood
-from .ssr import state_space_restriction
+from .training import state_space_restriction
 
 from cmath import log
 import numpy as np
