+++ mhn/training/state_space_restriction.pxd
